third_party/blink/web_tests/platform/mac-mac12-arm64/external/wpt/screen-orientation/lock-sandboxed-iframe-expected.txt

This is a testharness.js-based test.
FAIL Test without 'allow-orientation-lock' sandboxing directive assert_equals: screen.lockOrientation() throws a SecurityError expected "SecurityError" but got "error: SecurityError Failed to execute 'lock' on 'ScreenOrientation': The window is sandboxed and lacks the 'allow-orientation-lock' flag."
PASS Test with 'allow-orientation-lock' sandboxing directive
Harness: the test ran to completion.
